requirements.txt
altair==5.4.1
asn1crypto==1.5.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==2.2.1
cryptography==43.0.3
dotenv-0.0.5
gitdb==4.0.11
gitpython==3.1.43
idna==3.10
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.12.1
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pyarrow==18.0.0
pycparser==2.22
pydeck==0.9.1
pygments==2.18.0
pyjwt==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.1
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.12.3
snowflake-snowpark-python==1.24.0
sortedcontainers==2.4.0
streamlit==1.35.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==5.0.3
